+++ model/src/form/form-definition/index.ts
@@ -1,10 +1,13 @@
 import joi from 'joi'
 
+import { type FormDefinition } from '~/src/form/form-definition/types.js'
+
 /**
  * If an optional key is added, CURRENT_VERSION does not need to be incremented.
  * Only breaking changes will require an increment, as well as a migration script.
  */
 export const CURRENT_VERSION = 2
+
 const sectionsSchema = joi.object().keys({
   name: joi.string().required(),
   title: joi.string().required(),
@@ -263,6 +266,10 @@ const feeOptionSchema = joi
     }
   })
 
+/**
+ * Joi schema for `FormDefinition` interface
+ * @see {@link FormDefinition}
+ */
 export const Schema = joi
   .object()
   .required()
