+++ model/src/index.ts
@@ -1,4 +1,3 @@
-export * from '~/src/schema/index.js'
 export * from '~/src/data-model/index.js'
 export * from '~/src/utils/logger.js'
 export * from '~/src/form/index.js'
@@ -8,4 +7,4 @@ export * from '~/src/utils/helpers.js'
 export * from '~/src/migration/index.js'
 export type * from '~/src/components/types.js'
 export type * from '~/src/conditions/types.js'
-export type * from '~/src/data-model/types.js'
+export type * from '~/src/form/types.js'
