--- model/src/schema/index.ts
@@ -1,1 +1,0 @@
-export { Schema, componentSchema } from '~/src/schema/schema.js'
